+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/SendContact.kt
@@ -5,6 +5,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.requests.send.SendContact
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 
 suspend fun TelegramBot.sendContact(
     chatId: ChatIdentifier,
@@ -52,4 +53,34 @@ suspend fun TelegramBot.sendContact(
     replyMarkup: KeyboardMarkup? = null
 ) = sendContact(
     chat.id, contact, disableNotification, replyToMessageId, replyMarkup
+)
+
+suspend inline fun TelegramBot.reply(
+    to: Message,
+    phoneNumber: String,
+    firstName: String,
+    lastName: String? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = sendContact(
+    to.chat,
+    phoneNumber,
+    firstName,
+    lastName,
+    disableNotification,
+    to.messageId,
+    replyMarkup
+)
+
+suspend inline fun TelegramBot.reply(
+    to: Message,
+    contact: Contact,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = sendContact(
+    to.chat,
+    contact,
+    disableNotification,
+    to.messageId,
+    replyMarkup
 )