+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/media/SendAnimation.kt
@@ -7,7 +7,9 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.ParseMode
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.AnimationFile
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 
 suspend fun TelegramBot.sendAnimation(
     chatId: ChatIdentifier,
@@ -52,3 +54,78 @@ suspend fun TelegramBot.sendAnimation(
     chatId, animation.fileId, animation.thumb ?.fileId, text, parseMode, duration, width, height, disableNotification, replyToMessageId, replyMarkup
 )
 
+suspend fun TelegramBot.sendAnimation(
+    chat: Chat,
+    animation: InputFile,
+    thumb: InputFile? = null,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    width: Int? = null,
+    height: Int? = null,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = sendAnimation(chat.id, animation, thumb, text, parseMode, duration, width, height, disableNotification, replyToMessageId, replyMarkup)
+
+suspend fun TelegramBot.sendAnimation(
+    chat: Chat,
+    animation: AnimationFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    width: Int? = null,
+    height: Int? = null,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = sendAnimation(chat.id, animation, text, parseMode, duration, width, height, disableNotification, replyToMessageId, replyMarkup)
+
+suspend inline fun TelegramBot.replyWithAnimation(
+    to: Message,
+    animation: InputFile,
+    thumb: InputFile? = null,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    width: Int? = null,
+    height: Int? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = sendAnimation(
+    to.chat,
+    animation,
+    thumb,
+    text,
+    parseMode,
+    duration,
+    width,
+    height,
+    disableNotification,
+    to.messageId,
+    replyMarkup
+)
+
+suspend inline fun TelegramBot.replyWithAnimation(
+    to: Message,
+    animation: AnimationFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    width: Int? = null,
+    height: Int? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = sendAnimation(to.chat, animation, text, parseMode, duration, width, height, disableNotification, to.messageId, replyMarkup)
+
+suspend inline fun TelegramBot.reply(
+    to: Message,
+    animation: AnimationFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    width: Int? = null,
+    height: Int? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = replyWithAnimation(to, animation, text, parseMode, duration, width, height, disableNotification, replyMarkup)