+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/media/SendMediaGroup.kt
@@ -6,6 +6,7 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.InputMedia.MediaGroupMemberInputMedia
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 
 suspend fun TelegramBot.sendMediaGroup(
     chatId: ChatIdentifier,
@@ -26,3 +27,15 @@ suspend fun TelegramBot.sendMediaGroup(
 ) = sendMediaGroup(
     chat.id, media, disableNotification, replyToMessageId
 )
+
+suspend inline fun TelegramBot.replyWithMediaGroup(
+    to: Message,
+    media: List<MediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false
+) = sendMediaGroup(to.chat, media, disableNotification, to.messageId)
+
+suspend inline fun TelegramBot.reply(
+    to: Message,
+    media: List<MediaGroupMemberInputMedia>,
+    disableNotification: Boolean = false
+) = replyWithMediaGroup(to, media, disableNotification)