+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/send/media/SendVoice.kt
@@ -7,12 +7,14 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.ChatIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.ParseMode.ParseMode
 import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.chat.abstracts.Chat
 import com.github.insanusmokrassar.TelegramBotAPI.types.files.AudioFile
+import com.github.insanusmokrassar.TelegramBotAPI.types.files.VoiceFile
+import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.Message
 
 suspend fun TelegramBot.sendVoice(
     chatId: ChatIdentifier,
     voice: InputFile,
-    thumb: InputFile? = null,
     text: String? = null,
     parseMode: ParseMode? = null,
     duration: Long? = null,
@@ -23,7 +25,6 @@ suspend fun TelegramBot.sendVoice(
     SendVoice(
         chatId,
         voice,
-        thumb,
         text,
         parseMode,
         duration,
@@ -34,13 +35,62 @@ suspend fun TelegramBot.sendVoice(
 )
 
 suspend fun TelegramBot.sendVoice(
+    chat: Chat,
+    voice: InputFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = sendVoice(chat.id, voice, text, parseMode, duration, disableNotification, replyToMessageId, replyMarkup)
+
+suspend fun TelegramBot.sendVoice(
     chatId: ChatIdentifier,
-    voice: AudioFile,
+    voice: VoiceFile,
     text: String? = null,
     parseMode: ParseMode? = null,
     disableNotification: Boolean = false,
     replyToMessageId: MessageIdentifier? = null,
     replyMarkup: KeyboardMarkup? = null
 ) = sendVoice(
-    chatId, voice.fileId, voice.thumb ?.fileId, text, parseMode, voice.duration, disableNotification, replyToMessageId, replyMarkup
+    chatId, voice.fileId, text, parseMode, voice.duration, disableNotification, replyToMessageId, replyMarkup
 )
+
+suspend fun TelegramBot.sendVoice(
+    chat: Chat,
+    voice: VoiceFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    disableNotification: Boolean = false,
+    replyToMessageId: MessageIdentifier? = null,
+    replyMarkup: KeyboardMarkup? = null
+) = sendVoice(chat.id, voice, text, parseMode, disableNotification, replyToMessageId, replyMarkup)
+
+suspend inline fun TelegramBot.replyWithVoice(
+    to: Message,
+    voice: InputFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    duration: Long? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = sendVoice(to.chat, voice, text, parseMode, duration, disableNotification, to.messageId, replyMarkup)
+
+suspend inline fun TelegramBot.replyWithVoice(
+    to: Message,
+    voice: VoiceFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = sendVoice(to.chat, voice, text, parseMode, disableNotification, to.messageId, replyMarkup)
+
+suspend inline fun TelegramBot.reply(
+    to: Message,
+    voice: VoiceFile,
+    text: String? = null,
+    parseMode: ParseMode? = null,
+    disableNotification: Boolean = false,
+    replyMarkup: KeyboardMarkup? = null
+) = replyWithVoice(to, voice, text, parseMode, disableNotification, replyMarkup)