+++ CHANGELOG.md
@@ -46,9 +46,15 @@
 
 * `TelegramBotAPI-core`:
     * Now in forward info you can get `ForwardFromSupergroupInfo`
+    * **BREAKING CHANGE** `SendVoice` factory function has changed its signature: now it have now `thumb`
+    (according to the [documentation](https://core.telegram.org/bots/api#sendvoice))
 * `TelegramBotAPI-extensions-api`:
     * ALL REQUESTS EXECUTOR USAGES WERE REPLACED WITH `TelegramBot` TYPEALIAS. It should not bring any break changes
-
+    * Internal changes of `sendRegularPoll` and `sendQuizPoll` extensions
+    * Variable `defaultLivePeriodDelayMillis` now is public
+    * All `send` extensions for `TelegramBot` got their `reply` variations (issue [#144](https://github.com/InsanusMokrassar/TelegramBotAPI/issues/144))
+    * A lot of `send` extensions for `TelegramBot` got their variation with `Chat` instead of `ChatIdentifier`
+    
 ### 0.28.4
 
 * `Common`: